# api/openapi.yaml
openapi: 3.1.0
info:
  title: CloudOptix API
  version: "1.0.0"
  summary: >-
    Autonomous cloud cost intelligence and optimization: architecture-aware
    discovery, economics, recommendations, governed automation and an
    auditable execution trail, all tenant-isolated.
  description: |
    CloudOptix models a customer's AWS estate as a digital twin, prices it,
    finds savings, and — within policy — executes and rolls back changes
    itself. This document describes the whole HTTP API a client (the
    CloudOptix web app, a CI pipeline calling the cost compiler, or a partner
    integration) talks to.

    ## Authentication
    Every route except `/healthz`, `/readyz`, `/metrics`, `/openapi.yaml` and
    the `Onboarding` tag requires a credential: either an OIDC-issued bearer
    JWT (`Authorization: Bearer <token>`) or a service API key
    (`X-CloudOptix-Api-Key: <key>`). See the `BearerAuth` and `ApiKeyAuth`
    security schemes below. A worker or CI system uses a service token
    (issued out of band) presented the same way as a user's bearer token.

    ## Tenant scope
    An OIDC principal may belong to more than one tenant; which one a
    request is scoped to is resolved from the token's own tenant claim when
    present, or from the `X-CloudOptix-Tenant` header otherwise, with a
    membership check either way — a header can select among tenants a
    principal actually belongs to, never grant a new one. A service token
...
    ever names a tenant explicitly.

...

...

...

...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# YAML anchors for parameters and error responses shared across dozens of
# operations. A vendor extension at the document root (x-*, permitted
# anywhere per the OpenAPI 3.1 Specification Extensions rule) rather than a
# components/ entry, because these are resolved by the YAML parser itself —
# before any OpenAPI tooling sees the document — while a $ref into
# components/parameters would still repeat one line per operation for no
# benefit; the effect on the parsed document is identical to writing every
# field out longhand at every one of the ~115 operations below, just without
# ~115 chances for one of those copies to drift from the rest.
...
paths:
...
  /costs/summary:
    parameters: [*tenantHeader, *fromParam, *toParam]
    get:
      operationId: costs.summary
      tags: [Costs]
      summary: Get the cost-intelligence headline
      responses:
        "200":
          description: Summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CostSummary" }
        <<: *commonErrors
...
  /costs/explain:
    parameters: [*tenantHeader, *fromParam, *toParam]
    get:
      operationId: costs.explain
      tags: [Costs]
      summary: Explain a cost movement
      description: >-
        Decomposes the change between the requested period and the
        immediately preceding period of equal length into ranked
        contributors, answering "why did cost change" rather than merely
        reporting that it did.
      responses:
        "200":
          description: The explanation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CostExplanation" }
              example:
                current_total: { micros: 48210550000, amount: 48210.55, currency: USD, display: "$48,210.55" }
                baseline_total: { micros: 41190020000, amount: 41190.02, currency: USD, display: "$41,190.02" }
                delta: { micros: 7020530000, amount: 7020.53, currency: USD, display: "$7,020.53" }
                delta_pct: 0.1705
                contributors:
                  - dimension: service
                    key: AmazonRDS
                    delta: { micros: 4110000000, amount: 4110.00, currency: USD, display: "$4,110.00" }
                    share: 0.585
                    note: "A read replica was added in us-east-1 on the 12th."
                narrative: "Spend rose 17% week over week, driven mostly by a new RDS read replica."
        <<: *commonErrors
...
  /recommendations/analyze:
    parameters: [*tenantHeader, *idempotencyKey]
    post:
      operationId: recommendations.analyze
      tags: [Recommendations]
      summary: Run the full rule set and produce a fresh recommendation set
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnalyzeRequest" }
      responses:
        "202":
          description: Analysis result.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalyzeResult" }
        <<: *commonErrors
...
  /recommendations/summary:
    parameters: [*tenantHeader]
    get:
      operationId: recommendations.summary
      tags: [Recommendations]
      summary: Get the recommendations headline
      responses:
        "200":
          description: Summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendationSummary" }
        <<: *commonErrors
...
  /recommendations/{id}/explain:
    parameters: [*tenantHeader, *idPathParam]
    get:
      operationId: recommendations.explain
      tags: [Recommendations]
      summary: Get the full reasoning packet for one recommendation
      description: >-
        Evidence, confidence inputs, blast radius, risk factors, the policy
        decision and — when the recommendation has run before — similar
        historical outcomes.
      responses:
        "200":
          description: The explanation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecommendationExplanation" }
        <<: *commonErrors
  /recommendations/{id}/dismiss:
    parameters: [*tenantHeader, *idPathParam, *idempotencyKey]
    post:
      operationId: recommendations.dismiss
      tags: [Recommendations]
      summary: Dismiss a recommendation
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DismissRequest" }
      responses:
        "204": { description: Dismissed. }
        <<: *commonErrors
...
  /simulations/counterfactual:
    parameters: [*tenantHeader, *idempotencyKey]
    post:
      operationId: simulations.counterfactual
      tags: [Simulations]
      summary: Answer a what-if question
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Scenario" }
      responses:
        "200":
          description: The counterfactual answer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Counterfactual" }
              example:
                scenario: { type: add_cache, label: "Add ElastiCache in front of the checkout DB" }
                cost_delta: { micros: 640000000, amount: 640.00, currency: USD, display: "$640.00" }
                cost_delta_pct: 0.081
                risk: LOW
                confidence: 0.78
        <<: *commonErrors
...
  /compiler/compile:
    parameters: [*tenantHeader, *idempotencyKey]
    post:
      operationId: compiler.compile
      tags: [Cost Compiler]
      summary: Price an infrastructure-as-code change set
      description: >-
        Accepts the IaC artifact as a raw request body — a Terraform plan
        JSON, a CloudFormation template, a Kubernetes manifest, whatever
        `source` declares — rather than a JSON envelope, so a CI pipeline
        can pipe `terraform show -json plan.out` straight in.
      parameters:
        - { name: source, in: query, schema: { $ref: "#/components/schemas/SourceKind" }, example: terraform_plan }
        - { name: label, in: query, schema: { type: string }, example: "PR #482: add read replica" }
        - { name: region, in: query, schema: { type: string } }
        - { name: account_id, in: query, schema: { type: string } }
        - { name: environment, in: query, schema: { type: string } }
        - name: assume_*
          in: query
          schema: { type: number }
          description: >-
            Any `assume_<key>=<number>` query parameter overrides a
            usage-dependent resource's default assumption, e.g.
            `assume_lambda_invocations_per_month=40000000`.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
          text/plain:
            schema: { type: string }
      responses:
        "201":
          description: The priced change set.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompilationResult" }
        <<: *commonErrors
...
  /policies/simulate:
    parameters: [*tenantHeader]
    post:
      operationId: policies.simulate
      tags: [Policies]
      summary: Preview a proposed policy's impact on current recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Policy" }
      responses:
        "200":
          description: Impact preview.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicySimulation" }
        <<: *commonErrors
...
  /approvals:
    parameters: [*tenantHeader, *cursorParam, *limitParam, *sortParam, *orderParam]
    get:
      operationId: approvals.list
      tags: [Approvals]
      summary: List approval requests
      responses:
        "200":
          description: A page of approvals.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalPage" }
        <<: *commonErrors
    post:
      operationId: approvals.request
      tags: [Approvals]
      summary: Manually request an approval
      parameters: [*idempotencyKey]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalRequest" }
      responses:
        "201":
          description: Requested.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequest" }
        <<: *commonErrors
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        An OIDC-issued user JWT, an internal service token, or (in
        non-production deployments only) the fixed development static
        token. Validated per internal/infrastructure/auth: issuer, audience,
        expiry and signing algorithm are all checked, and `alg: none` is
        always rejected.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-CloudOptix-Api-Key
      description: A long-lived service API key, bound to exactly one tenant at issuance.
...
  schemas:
...
    RiskLevel:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH, CRITICAL]
...
    Money:
      type: object
      description: An exact monetary amount, stored as integer micros to avoid floating-point drift.
      required: [micros, currency, amount, display]
      properties:
        micros: { type: integer, format: int64, example: 4110000000, description: "Exact amount in 1/1,000,000ths of a unit." }
        currency: { type: string, example: USD }
        amount: { type: number, example: 4110.00, description: "Convenience float; micros is authoritative." }
        display: { type: string, example: "$4,110.00" }
    Period:
      type: object
      required: [start, end]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
...
    CostSeries:
      type: object
      properties:
        granularity: { type: string, enum: [hourly, daily, monthly] }
        points: { type: array, items: { $ref: "#/components/schemas/CostPoint" } }
        currency: { type: string }
...
    CostBreakdown:
      type: object
      properties:
        dimension: { type: string }
        period: { $ref: "#/components/schemas/Period" }
        total: { $ref: "#/components/schemas/Money" }
        items: { type: array, items: { $ref: "#/components/schemas/BreakdownItem" } }
    CostForecast:
      type: object
      properties:
        period: { $ref: "#/components/schemas/Period" }
        expected: { $ref: "#/components/schemas/Money" }
        low: { $ref: "#/components/schemas/Money" }
        high: { $ref: "#/components/schemas/Money" }
        method: { type: string, example: "holt_winters" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        based_on_days: { type: integer }
        note: { type: string }
    Contribution:
      type: object
      properties:
        dimension: { type: string }
        key: { type: string }
        delta: { $ref: "#/components/schemas/Money" }
        share: { type: number }
        note: { type: string }
...
    LinkedChange:
      type: object
      properties:
        kind: { type: string, enum: [execution, compilation, discovery_delta] }
        id: { type: string }
        label: { type: string }
        at: { type: string, format: date-time }
        cost_impact: { $ref: "#/components/schemas/Money" }
        correlation: { type: number }
    CostExplanation:
      type: object
      properties:
        current_period: { $ref: "#/components/schemas/Period" }
        baseline_period: { $ref: "#/components/schemas/Period" }
        current_total: { $ref: "#/components/schemas/Money" }
        baseline_total: { $ref: "#/components/schemas/Money" }
        delta: { $ref: "#/components/schemas/Money" }
        delta_pct: { type: number }
        contributors: { type: array, items: { $ref: "#/components/schemas/Contribution" } }
        narrative: { type: string }
        linked_changes: { type: array, items: { $ref: "#/components/schemas/LinkedChange" } }
    CostSummary:
      type: object
      properties:
        period: { $ref: "#/components/schemas/Period" }
        total: { $ref: "#/components/schemas/Money" }
        daily_average: { $ref: "#/components/schemas/Money" }
        month_to_date: { $ref: "#/components/schemas/Money" }
        prior_month: { $ref: "#/components/schemas/Money" }
        change_pct: { type: number }
        forecast: { $ref: "#/components/schemas/CostForecast" }
        by_service: { $ref: "#/components/schemas/CostBreakdown" }
        by_account: { $ref: "#/components/schemas/CostBreakdown" }
        by_environment: { $ref: "#/components/schemas/CostBreakdown" }
        by_application: { $ref: "#/components/schemas/CostBreakdown" }
        trend: { $ref: "#/components/schemas/CostSeries" }
        open_anomalies: { type: integer }
        last_ingested_at: { type: string, format: date-time }
        freshness: { type: string }

    # -------------------------------------------------------------
    # Economics / Cost SLOs
    # -------------------------------------------------------------
...
    AnalyzeRequest:
      type: object
      properties:
        account_ids: { type: array, items: { type: string } }
        environments: { type: array, items: { type: string } }
        categories: { type: array, items: { type: string } }
        rule_ids: { type: array, items: { type: string } }
        generate_narratives: { type: boolean, description: "Ask the LLM to write human explanations; never affects which recommendations exist or their scores." }
        async: { type: boolean }
    AnalyzeResult:
      type: object
      properties:
        run_id: { type: string }
        resources_analyzed: { type: integer }
        rules_evaluated: { type: integer }
        findings_produced: { type: integer }
        recommendations_created: { type: integer }
        superseded: { type: integer }
        mutually_exclusive_alternatives:
          type: integer
          description: >-
            How many of recommendations_created are alternatives within a
            conflict group and therefore excluded from total_monthly_saving.
        total_monthly_saving: { $ref: "#/components/schemas/Money" }
        total_annual_saving: { $ref: "#/components/schemas/Money" }
        auto_executable: { type: integer }
        requiring_approval: { type: integer }
        prohibited: { type: integer }
        duration_ms: { type: integer }
        warnings: { type: array, items: { type: string } }
...
    RecommendationSummary:
      type: object
      properties:
        open: { type: integer }
        total_monthly_saving: { $ref: "#/components/schemas/Money" }
        by_category: { type: object, additionalProperties: { type: integer } }
        saving_by_category: { type: object, additionalProperties: { $ref: "#/components/schemas/Money" } }
        by_risk: { type: object, additionalProperties: { type: integer } }
        auto_executable: { type: integer }
        awaiting_approval: { type: integer }
        mutually_exclusive_alternatives:
          type: integer
          description: >-
            How many open recommendations are alternatives to another one and
            therefore excluded from total_monthly_saving and
            saving_by_category. Counts include them; money does not.
...
    RecommendationExplanation:
      type: object
      properties:
        recommendation: { $ref: "#/components/schemas/Recommendation" }
        evidence: { type: array, items: { type: object, additionalProperties: true } }
        confidence_inputs: { type: array, items: { $ref: "#/components/schemas/ConfidenceInput" } }
        risk_factors: { type: array, items: { $ref: "#/components/schemas/RiskFactor" } }
        blast_radius: { $ref: "#/components/schemas/BlastRadius" }
        affected_nodes: { type: array, items: { $ref: "#/components/schemas/TwinNode" } }
        alternatives:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
          description: >-
            The other ways of fixing the same problem, ranked, so a reviewer
            sees every option and which one CloudOptix recommends. Empty when
            this recommendation competes with nothing.
        policy_decision: { $ref: "#/components/schemas/PolicyDecision" }
        calibration: { $ref: "#/components/schemas/RuleCalibration" }
        rollback_summary: { type: string }
        narrative: { type: string }
        similar_outcomes: { type: array, items: { $ref: "#/components/schemas/Outcome" } }
...
    DismissRequest:
      type: object
      properties:
        reason: { type: string }
...
    Scenario:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [traffic_change, platform_change, database_change, add_cache, remove_nat, add_vpc_endpoint, spot_adoption, region_change, commitment_purchase, storage_class_change, replica_change, custom]
        label: { type: string }
        parameters: { type: object, additionalProperties: true }
        scope_id: { type: string }
...
    Counterfactual:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        scenario: { $ref: "#/components/schemas/Scenario" }
        question: { type: string }
        current_state: { $ref: "#/components/schemas/StateProjection" }
        proposed_state: { $ref: "#/components/schemas/StateProjection" }
        cost_delta: { $ref: "#/components/schemas/Money" }
        cost_delta_pct: { type: number }
        annual_cost_delta: { $ref: "#/components/schemas/Money" }
        performance_delta: { type: string }
        reliability_delta: { type: string }
        security_delta: { type: string }
        risk: { $ref: "#/components/schemas/RiskLevel" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        assumptions: { type: array, items: { $ref: "#/components/schemas/Assumption" } }
        caveats: { type: array, items: { type: string } }
        narrative: { type: string }
        computed_at: { type: string, format: date-time }
    SourceKind:
      type: string
      enum: [terraform_plan, terraform_hcl, cloudformation, kubernetes_manifest, helm_release, live_topology]
...
    CompilationResult:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        source: { $ref: "#/components/schemas/SourceKind" }
        label: { type: string }
        changes: { type: array, items: { $ref: "#/components/schemas/PricedChange" } }
        baseline_monthly: { $ref: "#/components/schemas/Money" }
        projected_monthly: { $ref: "#/components/schemas/Money" }
        monthly_delta: { $ref: "#/components/schemas/Money" }
        annual_delta: { $ref: "#/components/schemas/Money" }
        delta_pct: { type: number }
...
    Policy:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: integer }
        rules: { type: array, items: { $ref: "#/components/schemas/PolicyRule" } }
        default_effect: { type: string, enum: [auto_execute, require_approval, prohibit, advisory], description: "Validation refuses auto_execute here — an unmatched action is by definition unreviewed." }
        enabled: { type: boolean }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        activated_at: { type: string, format: date-time }
        checksum: { type: string }
...
    PolicySimulation:
      type: object
      properties:
        policy_name: { type: string }
        evaluated: { type: integer }
        auto_execute: { type: integer }
        require_approval: { type: integer }
        prohibited: { type: integer }
        advisory: { type: integer }
        changes: { type: array, items: { $ref: "#/components/schemas/PolicySimulationChange" } }
        auto_executable_saving: { $ref: "#/components/schemas/Money" }
        warnings: { type: array, items: { type: string } }
...
    ApprovalRequest:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        subject_kind: { type: string, enum: [execution_plan, policy, aws_connection, commitment_purchase] }
        subject_id: { type: string }
        title: { type: string }
        summary: { type: string }
        context: { $ref: "#/components/schemas/ApprovalContext" }
        policy_decision_id: { type: string }
        required_roles: { type: array, items: { type: string } }
        min_approvals: { type: integer }
        require_distinct_approver: { type: boolean }
        state: { type: string, enum: [pending, approved, rejected, expired, withdrawn] }
        responses: { type: array, items: { $ref: "#/components/schemas/ApprovalResponse" } }
        requested_by: { type: string }
        requested_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        decided_at: { type: string, format: date-time }
        execute_after: { type: string, format: date-time }
    ApprovalPage:
      allOf:
        - $ref: "#/components/schemas/PageEnvelope"
        - type: object
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/ApprovalRequest" } }
...